refactored statistics reporting
