+++ quickwit-cli/Cargo.toml
@@ -16,3 +16,4 @@ tracing = '0.1'
 tracing-subscriber = "0.2"
 tokio = { version = "1", features = ["full"] }
 tokio-util = { version = "0.6", features = ["full"] }
+crossterm = "0.19.0"
