+++ quickwit-core/Cargo.toml
@@ -24,8 +24,6 @@ serde_json = "1.0"
 tracing = '0.1'
 tracing-subscriber = "0.2"
 tokio-stream = "0.1.6"
-crossterm = "0.19.0"
-once_cell = "1"
 
 [dev-dependencies]
 tempfile = '3'
