+++ quickwit-core/src/indexing/split.rs
@@ -38,7 +38,6 @@ use tracing::{info, warn};
 use uuid::Uuid;
 
 use super::IndexDataParams;
-use super::INDEXING_STATISTICS;
 
 pub const MAX_DOC_PER_SPLIT: usize = if cfg!(test) { 100 } else { 5_000_000 };
 
@@ -169,19 +168,12 @@ impl Split {
         self.metastore
             .stage_split(&self.index_id, self.metadata.clone())
             .await?;
-
-        INDEXING_STATISTICS.num_staged_splits.inc();
         Ok(self.id.to_string())
     }
 
     /// Upload all split artifacts using the storage.
     pub async fn upload(&self) -> anyhow::Result<Manifest> {
         let manifest = put_to_storage(&*self.storage, self).await?;
-
-        INDEXING_STATISTICS.num_uploaded_splits.inc();
-        INDEXING_STATISTICS
-            .total_size_splits
-            .add(manifest.split_size_in_bytes as usize);
         Ok(manifest)
     }
 
@@ -190,8 +182,6 @@ impl Split {
         self.metastore
             .publish_split(&self.index_uri, &self.id.to_string())
             .await?;
-
-        INDEXING_STATISTICS.num_published_splits.inc();
         Ok(())
     }
 }
